// src/app/shared/models/assessment.model.ts
/*
 * Copyright 2025 TASKEXPOSURE
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Task } from './task.model.ts';


export type SignalStrength = 'low' | 'moderate' | 'high';

export interface RoadmapAction {
  title: string;
  description: string;
  mechanism: string;
  impact: string;
}

export interface ClarifyingQuestion {
  prompt: string;
  options: string[];
}

export interface AssessmentData {
  tasks: Task[];
  aiCapabilities: string[];
  aiLimitations: string[];
  signalStrength: SignalStrength;
  riskScore: number;
  roadmapActions: RoadmapAction[];
}